Bug fix, better pose recognition, mining config, more animations
Fixed issue #1 where the "die" animation caused actions (besides teleporting) to break if you tried to cancel the action.
Changed the way the player's poses are identified to allow for more stable movement animation overrides.
Added mining config to the skilling section.
Added a couple new animations to be used in the configs.
Reset the player's idle pose if they disable the plugin.